+++ final/resource.h
@@ -5,8 +5,24 @@
 #define IDD_DIALOG_AUTH 1001
 
 // Menu IDs
-#define IDM_EXIT          101
-#define IDM_SHOW_DEVICES  102
-#define IDM_TOGGLE_LOGGING 103
+#define IDM_EXIT                101
+#define IDM_SHOW_DEVICES        102
+#define IDM_TOGGLE_LOGGING      103
+#define IDM_SETTINGS            104
+#define IDM_ABOUT               105
+#define IDM_TRUST_DEVICE        106
+#define IDM_UNTRUST_DEVICE      107
+#define IDM_DEVICE_DETAILS      108
+
+// Control IDs
+#define IDC_DEVICE_LIST         201
+#define IDC_STATUS_BAR          202
+#define IDC_CB_ENABLE_LOGGING   203
+#define IDC_CB_BLOCK_SUSPICIOUS 204
+#define IDC_CB_STARTUP          205
+#define IDC_EDIT_KEYWORD        206
+#define IDC_BTN_ADD_KEYWORD     207
+#define IDC_LIST_BLACKLIST      208
+#define IDC_BTN_REMOVE_KEYWORD  209
 
 #endif // RESOURCE_H